requirements.txt
streamlit==1.31.0
numpy==1.26.4
opencv-python-headless==4.9.0.80
Pillow==10.3.0
requests==2.32.3
